+++ src/main/java/kakaotech/bootcamp/respec/specranking/chatconsumer/domain/chat/adapter/in/exception/ChatEventException.java
@@ -1,0 +1,11 @@
+package kakaotech.bootcamp.respec.specranking.chatconsumer.domain.chat.adapter.in.exception;
+
+public class ChatEventException extends RuntimeException {
+    public ChatEventException(String message) {
+        super(message);
+    }
+
+    public ChatEventException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
